+++ tests/io/test_cdips.py
@@ -34,6 +34,7 @@ def test_read_cdips():
     for ext in exts:
         lc = read_cdips_lightcurve(url, flux_column=ext)
         assert type(lc).__name__ == "TessLightCurve"
+        assert lc.meta["FLUX_ORIGIN"] == ext.lower()
         # Are `time` and `flux` consistent with the FITS file?
         assert_array_equal(f[1].data['TMID_BJD'][lc.meta['QUALITY_MASK']],
                            lc.time.value)